+++ src/components/banner/banner.js
@@ -18,8 +18,10 @@ export default function main(_) {
      * （该处理需在loading结束之前处理）
      */
     (() => {
-        let mainHeader = $('.main-header');
+        let mainHeader = $('#main-header');
         let topImg, bgImg, height;
+
+        // 设置图片
         if (_.__status.pageType === 'home') {
             topImg = _.__config.banner.home.background.length > 0
                 ? _.__config.banner.home.background : [homeTopBg];
@@ -33,15 +35,36 @@ export default function main(_) {
             $('#cb_post_title_url').addClass('post-del-title');
         }
 
-        topImg.length > 0 ?
-            (topImg.length > 1 ? bgImg = topImg[_.__tools.randomNum(0, topImg.length - 1)] : bgImg = topImg[0])
-            : bgImg = "";
-
-        mainHeader.css({
-            'background': '#222 url(\'' + encodeURI(bgImg) + '\')  center center no-repeat',
-            'background-size': 'cover'
-        });
+        // 设置高度
         if (height) mainHeader.css('height', height);
+
+        // banner动效
+        if (_.__config.animate.bannerImages.enable) {
+            // 开启图片自动切换
+            import(/* webpackChunkName: "bannerImages" */ '../bannerImages/bannerImages').then(module => {
+                let bannerImages = module.default;
+                bannerImages(
+                    'main-header',
+                    topImg,
+                    _.__config.animate.bannerImages.options.itemNum,
+                    _.__config.animate.bannerImages.options.time,
+                    _.__config.animate.bannerImages.options.sort,
+                    _.__config.animate.bannerImages.options.current < 0
+                        ? _.__tools.randomNum(0, topImg.length - 1)
+                        : _.__config.animate.bannerImages.options.current
+                );
+            });
+        } else {
+            // 随机指定一个图片
+            topImg.length > 0 ?
+                (topImg.length > 1 ? bgImg = topImg[_.__tools.randomNum(0, topImg.length - 1)] : bgImg = topImg[0])
+                : bgImg = "";
+
+            mainHeader.css({
+                'background': '#222 url(\'' + encodeURI(bgImg) + '\')  center center no-repeat',
+                'background-size': 'cover'
+            });
+        }
     })();
 
     // 添加事件监听